// src/forgetrustee/frontend/src/types/generated/dataTrusteeModelResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DataTrusteeModelBusinessResponseDto } from './dataTrusteeModelBusinessResponseDto';
import type { DataTrusteeModelCoreResponseDto } from './dataTrusteeModelCoreResponseDto';
import type { DataTrusteeModelDataResponseDto } from './dataTrusteeModelDataResponseDto';
import type { DataTrusteeModelLegalResponseDto } from './dataTrusteeModelLegalResponseDto';
import type { DataTrusteeModelObjectivesResponseDto } from './dataTrusteeModelObjectivesResponseDto';
import type { DataTrusteeModelPresentationDto } from './dataTrusteeModelPresentationDto';

export interface DataTrusteeModelResponseDto {
  core?: DataTrusteeModelCoreResponseDto;
  data?: DataTrusteeModelDataResponseDto;
  legal?: DataTrusteeModelLegalResponseDto;
  objectives?: DataTrusteeModelObjectivesResponseDto;
  business?: DataTrusteeModelBusinessResponseDto;
  presentation?: DataTrusteeModelPresentationDto;
}