// src/forgetrustee/frontend/src/types/generated/dataTrusteeWizardBuildResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DataTrusteeModelResponseDto } from './dataTrusteeModelResponseDto';
import type { DataTrusteeWizardRequestDto } from './dataTrusteeWizardRequestDto';
import type { ShaclReportDTO } from './shaclReportDTO';

export interface DataTrusteeWizardBuildResponseDto {
  request?: DataTrusteeWizardRequestDto;
  dataTrusteeModel?: DataTrusteeModelResponseDto;
  ttl?: string;
  shaclReport?: ShaclReportDTO;
  executiveSummary?: string;
}